Workaround to make bootsnap work with CircleCI
There's a bug in bunbler or bootsnap or CircleCI. If you have bundler
1.16.0 and are using CircleCI with bootsnap, circle fails to find gems
(I think). The bug is documented here:
https://github.com/Shopify/bootsnap/issues/103

My workaround for now is to essentially disable bootsnap when on
CircleCI by looking for the CIRCLECI environment variable and using that
to set load_path_cache and autoload_paths_cache to false. All other
environments set it to true.